Changed clientGRPC to private method and moved to the package pkg/grpc
