+++ pkg/grpc/serverGRPC_test.go
@@ -3,7 +3,6 @@ package grpc
 import (
 	"context"
 	"github.com/stretchr/testify/require"
-	"monorepa/service/items"
 
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/test/bufconn"
@@ -24,12 +23,12 @@ func TestGrpcServiceServer(t *testing.T) {
 		{Title: "XVLBZGBAICMRAJWW", Description: "  fdzdgrxomvt ler"},
 	}
 
-	getItems, _ := client.GetItems(context.Background(), "I")
+	getItems, _ := client.getItems(context.Background(), "I")
 
 	require.Equal(t, x[0].Title, getItems[0].Title, "need to fix test program")
 	require.Equal(t, x[0].Description, getItems[0].Description, "need to fix test program")
 
-	_, err := client.GetItems(context.Background(), "")
+	_, err := client.getItems(context.Background(), "")
 	require.NotNil(t, err, "grpc should return err invalid username")
 }
 
@@ -41,7 +40,7 @@ func serveBufconn(ln *bufconn.Listener, data storage.ItemService) {
 	_ = s.Serve(ln)
 }
 
-func makeBufconnClient(ln *bufconn.Listener) *items.GRPCClient {
+func makeBufconnClient(ln *bufconn.Listener) *gRPCClient {
 
 	makeBufDialer := func(ln *bufconn.Listener) func(context.Context, string) (net.Conn, error) {
 		return func(context.Context, string) (net.Conn, error) {
@@ -49,7 +48,7 @@ func makeBufconnClient(ln *bufconn.Listener) *items.GRPCClient {
 		}
 	}
 
-	client := &items.GRPCClient{}
+	client := &gRPCClient{}
 
 	conn, _ := grpc.DialContext(
 		context.Background(),
@@ -58,7 +57,7 @@ func makeBufconnClient(ln *bufconn.Listener) *items.GRPCClient {
 		grpc.WithInsecure(),
 	)
 
-	client.Client = monorepa.NewGrpcServiceClient(conn)
+	client.client = monorepa.NewGrpcServiceClient(conn)
 
 	return client
 }
